requirements.txt
arc4==0.0.4
bitstring==3.1.9
hexdump==3.3
pycryptodome==3.12.0
WMI==1.5.1